Refactor Filament resource for Product: Add `filament-adjacency-list` dependency
This commit adds the `saade/filament-adjacency-list` package as a dependency in the `composer.json` file. This package provides functionality for managing hierarchical data structures in Filament resources. By using the `filament-adjacency-list` package, we can easily implement features like nested categories for products, improving the organization and management of product data in the application.